Remove val_splitter file
